refactor: muda o local do código da page de detalhes de uma pessoa para o app management #129
