+++ apps/management-app/src/components/ui/avatar.tsx
