+++ apps/management-app/src/lib/ui/DocumentCategoriesCard.tsx
